requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
django-environ